requirements.txt
discord.py==1.6.0
SQLAlchemy==1.3.22
# psycopg2==2.8.6
psycopg2-binary==2.8.6
